+++ app/src/main/java/com/ragdroid/rxify/codelab/ChillPresenter.java
@@ -1,8 +1,7 @@
 package com.ragdroid.rxify.codelab;
 
+import com.ragdroid.rxify.codelab.presenter.BaseCLPresenter;
 import com.ragdroid.rxify.core.BaseSchedulerProvider;
-
-import java.util.Arrays;
 
 import javax.inject.Inject;
 
@@ -13,7 +12,12 @@ import io.reactivex.disposables.Disposable;
  * Created by garimajain on 15/01/17.
  */
 
-public class ChillPresenter extends BaseCLPresenter<Integer> implements CodeLabContract.Presenter {
+public class ChillPresenter extends BaseCLPresenter<Object> implements CodeLabContract.Presenter {
+
+    //Input
+    Observable<Integer> inputValues = Observable.range(0,10);
+
+    //TODO Print all even numbers
 
     @Inject
     public ChillPresenter(BaseSchedulerProvider provider) {
@@ -22,7 +26,7 @@ public class ChillPresenter extends BaseCLPresenter<Integer> implements CodeLabC
 
     @Override
     protected Disposable getDisposable() {
-        return Observable.fromIterable(Arrays.asList(1, 2, 3, 4, 5, 6))
+        return Observable.empty()
                 .compose(lazyTransformer)
                 .subscribe(next, error, complete);
     }
