+++ app/src/main/java/com/ragdroid/rxify/codelab/presenter/FromPresenter.java
@@ -1,6 +1,9 @@
-package com.ragdroid.rxify.codelab;
+package com.ragdroid.rxify.codelab.presenter;
 
+import com.ragdroid.rxify.codelab.CodeLabContract;
 import com.ragdroid.rxify.core.BaseSchedulerProvider;
+
+import java.util.Arrays;
 
 import javax.inject.Inject;
 
@@ -11,16 +14,16 @@ import io.reactivex.disposables.Disposable;
  * Created by garimajain on 15/01/17.
  */
 
-public class ErrorPresenter extends BaseCLPresenter<Object> implements CodeLabContract.Presenter {
+public class FromPresenter extends BaseCLPresenter<Integer> implements CodeLabContract.Presenter {
 
     @Inject
-    public ErrorPresenter(BaseSchedulerProvider provider) {
+    public FromPresenter(BaseSchedulerProvider provider) {
         super(provider);
     }
 
     @Override
     protected Disposable getDisposable() {
-        return Observable.error(new RuntimeException("BOOM!"))
+        return Observable.fromIterable(Arrays.asList(1, 2, 3, 4, 5, 6))
                 .compose(lazyTransformer)
                 .subscribe(next, error, complete);
     }
