+++ app/src/main/java/com/ragdroid/rxify/codelab/presenter/IntervalRangePresenter.java
@@ -1,6 +1,9 @@
-package com.ragdroid.rxify.codelab;
+package com.ragdroid.rxify.codelab.presenter;
 
+import com.ragdroid.rxify.codelab.CodeLabContract;
 import com.ragdroid.rxify.core.BaseSchedulerProvider;
+
+import java.util.concurrent.TimeUnit;
 
 import javax.inject.Inject;
 
@@ -11,16 +14,16 @@ import io.reactivex.disposables.Disposable;
  * Created by garimajain on 15/01/17.
  */
 
-public class RangePresenter extends BaseCLPresenter<Integer> implements CodeLabContract.Presenter {
+public class IntervalRangePresenter extends BaseCLPresenter<Long> implements CodeLabContract.Presenter {
 
     @Inject
-    public RangePresenter(BaseSchedulerProvider provider) {
+    public IntervalRangePresenter(BaseSchedulerProvider provider) {
         super(provider);
     }
 
     @Override
     protected Disposable getDisposable() {
-        return Observable.range(1, 10)
+        return Observable.intervalRange(20, 5, 1000, 1000, TimeUnit.MILLISECONDS)
                 .compose(lazyTransformer)
                 .subscribe(next, error, complete);
     }
