+++ app/src/main/java/com/ragdroid/rxify/codelab/presenter/JustPresenter.java
@@ -1,8 +1,7 @@
-package com.ragdroid.rxify.codelab;
+package com.ragdroid.rxify.codelab.presenter;
 
+import com.ragdroid.rxify.codelab.CodeLabContract;
 import com.ragdroid.rxify.core.BaseSchedulerProvider;
-
-import java.util.concurrent.TimeUnit;
 
 import javax.inject.Inject;
 
@@ -13,16 +12,16 @@ import io.reactivex.disposables.Disposable;
  * Created by garimajain on 15/01/17.
  */
 
-public class TimerPresenter extends BaseCLPresenter<Long> implements CodeLabContract.Presenter {
+public class JustPresenter extends BaseCLPresenter<String> implements CodeLabContract.Presenter {
 
     @Inject
-    public TimerPresenter(BaseSchedulerProvider provider) {
+    public JustPresenter(BaseSchedulerProvider provider) {
         super(provider);
     }
 
     @Override
     protected Disposable getDisposable() {
-        return Observable.timer(1, TimeUnit.SECONDS)
+        return Observable.just("Hello World!")
                 .compose(lazyTransformer)
                 .subscribe(next, error, complete);
     }
