+++ app/src/main/java/com/ragdroid/rxify/codelab/presenter/NeverPresenter.java
@@ -1,5 +1,6 @@
-package com.ragdroid.rxify.codelab;
+package com.ragdroid.rxify.codelab.presenter;
 
+import com.ragdroid.rxify.codelab.CodeLabContract;
 import com.ragdroid.rxify.core.BaseSchedulerProvider;
 
 import javax.inject.Inject;
