+++ app/src/main/java/com/ragdroid/rxify/codelab/presenter/TimerPresenter.java
@@ -1,8 +1,9 @@
-package com.ragdroid.rxify.codelab;
+package com.ragdroid.rxify.codelab.presenter;
 
+import com.ragdroid.rxify.codelab.CodeLabContract;
 import com.ragdroid.rxify.core.BaseSchedulerProvider;
 
-import java.util.Arrays;
+import java.util.concurrent.TimeUnit;
 
 import javax.inject.Inject;
 
@@ -13,16 +14,16 @@ import io.reactivex.disposables.Disposable;
  * Created by garimajain on 15/01/17.
  */
 
-public class FromPresenter extends BaseCLPresenter<Integer> implements CodeLabContract.Presenter {
+public class TimerPresenter extends BaseCLPresenter<Long> implements CodeLabContract.Presenter {
 
     @Inject
-    public FromPresenter(BaseSchedulerProvider provider) {
+    public TimerPresenter(BaseSchedulerProvider provider) {
         super(provider);
     }
 
     @Override
     protected Disposable getDisposable() {
-        return Observable.fromIterable(Arrays.asList(1, 2, 3, 4, 5, 6))
+        return Observable.timer(1, TimeUnit.SECONDS)
                 .compose(lazyTransformer)
                 .subscribe(next, error, complete);
     }
